+++ src/index.tsx
@@ -3,6 +3,27 @@ export {PieChart} from './PieChart';
 export {LineChart} from './LineChart';
 export {LineChartBicolor} from './LineChart/LineChartBicolor';
 export {PopulationPyramid} from './PopulationPyramid';
-export {PopnPyramidItem} from './PopulationPyramid/types';
+
+export {
+  barDataItem,
+  stackDataItem,
+  BarChartPropsType,
+  StackedBarChartPropsType,
+} from './BarChart/types';
+
+export {pieDataItem, PieChartPropsType} from './PieChart/types';
+
+export {
+  lineDataItem,
+  bicolorLineDataItem,
+  LineChartPropsType,
+  LineChartBicolorPropsType,
+} from './LineChart/types';
+
+export {
+  popnPyramidDataItem,
+  PopulationPyramidPropsType,
+} from './PopulationPyramid/types';
+
 export {chartTypes, yAxisSides, ruleTypes} from './utils/constants';
 export {CurveType, EdgePosition} from './utils/types';